pruned release 49.0.2623.112

+++ src/content/renderer/media/webmediaplayer_ms.cc
@@ -4,11 +4,14 @@
 
 #include "content/renderer/media/webmediaplayer_ms.h"
 
+#include <stddef.h>
 #include <limits>
 #include <string>
+#include <utility>
 
 #include "base/bind.h"
 #include "base/callback.h"
+#include "build/build_config.h"
 #include "cc/blink/context_provider_web_context.h"
 #include "cc/blink/web_layer_impl.h"
 #include "cc/layers/video_frame_provider_client_impl.h"
@@ -47,22 +50,26 @@ WebMediaPlayerMS::WebMediaPlayerMS(
     media::GpuVideoAcceleratorFactories* gpu_factories,
     const blink::WebString& sink_id,
     const blink::WebSecurityOrigin& security_origin)
-    : frame_(frame),
+    : RenderFrameObserver(RenderFrame::FromWebFrame(frame)),
+      frame_(frame),
       network_state_(WebMediaPlayer::NetworkStateEmpty),
       ready_state_(WebMediaPlayer::ReadyStateHaveNothing),
       buffered_(static_cast<size_t>(0)),
       client_(client),
       delegate_(delegate),
       paused_(true),
+      render_frame_suspended_(false),
       received_first_frame_(false),
       media_log_(media_log),
-      renderer_factory_(factory.Pass()),
+      renderer_factory_(std::move(factory)),
       media_task_runner_(media_task_runner),
       worker_task_runner_(worker_task_runner),
       gpu_factories_(gpu_factories),
       compositor_task_runner_(compositor_task_runner),
       initial_audio_output_device_id_(sink_id.utf8()),
-      initial_security_origin_(security_origin) {
+      initial_security_origin_(security_origin.isNull()
+                                   ? url::Origin()
+                                   : url::Origin(security_origin)) {
   DVLOG(1) << __FUNCTION__;
   DCHECK(client);
   media_log_->AddEvent(
@@ -73,7 +80,7 @@ WebMediaPlayerMS::~WebMediaPlayerMS() {
   DVLOG(1) << __FUNCTION__;
   DCHECK(thread_checker_.CalledOnValidThread());
 
-  if (compositor_)
+  if (compositor_ && !compositor_task_runner_->BelongsToCurrentThread())
     compositor_task_runner_->DeleteSoon(FROM_HERE, compositor_.release());
 
   get_client()->setWebLayer(nullptr);
@@ -106,7 +113,7 @@ void WebMediaPlayerMS::load(LoadType load_type,
 
   SetNetworkState(WebMediaPlayer::NetworkStateLoading);
   SetReadyState(WebMediaPlayer::ReadyStateHaveNothing);
-  media_log_->AddEvent(media_log_->CreateLoadEvent(url.spec()));
+  media_log_->AddEvent(media_log_->CreateLoadEvent(url.string().utf8()));
 
   video_frame_provider_ = renderer_factory_->GetVideoFrameProvider(
       url,
@@ -117,9 +124,12 @@ void WebMediaPlayerMS::load(LoadType load_type,
       gpu_factories_);
 
   RenderFrame* const frame = RenderFrame::FromWebFrame(frame_);
-  audio_renderer_ = renderer_factory_->GetAudioRenderer(
-      url, frame->GetRoutingID(), initial_audio_output_device_id_,
-      initial_security_origin_);
+
+  if (frame) {
+    audio_renderer_ = renderer_factory_->GetAudioRenderer(
+        url, frame->GetRoutingID(), initial_audio_output_device_id_,
+        initial_security_origin_);
+  }
 
   if (!video_frame_provider_ && !audio_renderer_) {
     SetNetworkState(WebMediaPlayer::NetworkStateNetworkError);
@@ -350,6 +360,28 @@ unsigned WebMediaPlayerMS::videoDecodedByteCount() const {
   return 0;
 }
 
+void WebMediaPlayerMS::WasHidden() {
+#if defined(OS_ANDROID)
+  DCHECK(thread_checker_.CalledOnValidThread());
+  DCHECK(!render_frame_suspended_);
+
+  // Method called when the RenderFrame is sent to background and suspended
+  // (android). Substitute the displayed VideoFrame with a copy to avoid
+  // holding on to it unnecessarily.
+  render_frame_suspended_=true;
+  if (!paused_)
+    compositor_->ReplaceCurrentFrameWithACopy();
+#endif  // defined(OS_ANDROID)
+}
+
+void WebMediaPlayerMS::WasShown() {
+#if defined(OS_ANDROID)
+  DCHECK(thread_checker_.CalledOnValidThread());
+
+  render_frame_suspended_ = false;
+#endif  // defined(OS_ANDROID)
+}
+
 bool WebMediaPlayerMS::copyVideoTextureToPlatformTexture(
     blink::WebGraphicsContext3D* web_graphics_context,
     unsigned int texture,
@@ -382,6 +414,9 @@ void WebMediaPlayerMS::OnFrameAvailable(
     const scoped_refptr<media::VideoFrame>& frame) {
   DVLOG(3) << __FUNCTION__;
   DCHECK(thread_checker_.CalledOnValidThread());
+
+  if (render_frame_suspended_)
+    return;
 
   base::TimeTicks render_time;
   if (frame->metadata()->GetTimeTicks(media::VideoFrameMetadata::REFERENCE_TIME,
@@ -452,4 +487,5 @@ void WebMediaPlayerMS::ResetCanvasCache() {
   DCHECK(thread_checker_.CalledOnValidThread());
   video_renderer_.ResetCache();
 }
+
 }  // namespace content
